created and consumed cities context api
