+++ App.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
+import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import HomePage from "./src/pages/HomePage";
 import Product from "./src/pages/Product";
 import Pricing from "./src/pages/Pricing";
@@ -12,56 +12,33 @@ import Form from "./src/components/Form";
 import CityList from "./src/components/CityList";
 import City from "./src/components/City";
 import CountryList from "./src/components/CountryList";
+import { CitiesProvider } from "./src/context/CitiesContext";
 
 function App() {
-
-  const [cities, setCities] = useState([])
-  const [isLoading, setIstLoading] = useState(false)
-  const [error, setError] = useState("")
-
-  const BASE_URL = "http://localhost:8000"
-
-  useEffect(() => {
-    setIstLoading(true)
-    const fetchCities =  async () => {
-      try {
-        const res = await fetch(`${BASE_URL}/cities`)
-        if(res.ok) {
-          const data = await res.json()
-          
-          setCities(data)
-          setIstLoading(false)
-        }
-      } catch (error) {
-        setError(error.message)
-      }finally {
-        setIstLoading(false)
-      }
-    }
-    fetchCities()
-  },[])
   return (
     <div>
-      <BrowserRouter>
-        {/* <Header /> */}
-        <main>
-          <Routes>
-            <Route path="/" element={<HomePage />} />
-            <Route path="products" element={<Product />} />
-            <Route path="pricing" element={<Pricing />} />
-            <Route path="login" element={<Login />} />
-            <Route path="app" element={<AppLayout />}>
-              <Route index element={<CityList cities={cities} isLoading={isLoading} error={error}/>} />
-              <Route path="cities" element={<CityList cities={cities} isLoading={isLoading} error={error}/>} />
-              <Route path="cities/:id" element={<City />}/>
-              <Route path="countries" element={<CountryList cities={cities} isLoading={isLoading} error={error}/>} />
-              <Route path="form" element={<Form />} />
-            </Route>
-            <Route path="*" element={<PageNotFound />} />
-          </Routes>
-        </main>
-        {/* <Footer /> */}
-      </BrowserRouter>
+      <CitiesProvider>
+        <BrowserRouter>
+          {/* <Header /> */}
+          <main>
+            <Routes>
+              <Route path="/" element={<HomePage />} />
+              <Route path="products" element={<Product />} />
+              <Route path="pricing" element={<Pricing />} />
+              <Route path="login" element={<Login />} />
+              <Route path="app" element={<AppLayout />}>
+                <Route index element={<Navigate replace to="cities" />} />
+                <Route path="cities" element={<CityList />} />
+                <Route path="cities/:id" element={<City />} />
+                <Route path="countries" element={<CountryList />} />
+                <Route path="form" element={<Form />} />
+              </Route>
+              <Route path="*" element={<PageNotFound />} />
+            </Routes>
+          </main>
+          {/* <Footer /> */}
+        </BrowserRouter>
+      </CitiesProvider>
     </div>
   );
 }
