+++ src/components/CountryList.jsx
@@ -3,8 +3,10 @@ import styles from "./CountryList.module.css";
 import Spinner from "./Spinner";
 import Message from "./Message";
 import CountryItem from "./CountryItem";
+import { useCities } from "../context/CitiesContext";
 
-const CountriesList = ({ cities, isLoading, error }) => {
+const CountriesList = () => {
+  const {cities,isLoading,error} = useCities()
   if (isLoading) return <Spinner />;
   if (error) return <p>{error.message}</p>;
   if (!cities.length)
